Set type of requested_frames of all mixer callbacks to int
The function pointer expects an int so using anything else is either misleading and an implicit cast at best or possibly undefined behavior at worst.
In edge cases, the callbacks using uint16_t can also overflow so this change prevents that bug as well.

+++ src/hardware/innovation.cpp
@@ -221,7 +221,7 @@ bool Innovation::MaybeRenderFrame(float &frame)
 	return frame_is_ready;
 }
 
-void Innovation::AudioCallback(const uint16_t requested_frames)
+void Innovation::AudioCallback(const int requested_frames)
 {
 	assert(channel);
